add "local" dir for config, try to improve code readability
fix typo in readme

include local config independently of working dir
