+++ .gitignore
@@ -1,1 +1,1 @@
-/config.php
+local
