+++ local-template/local.js
@@ -1,0 +1,9 @@
+// This JS is included in index.php
+
+// Example: Fix height of iframe (adapt to your installation).
+
+//let iFrame = window.parent.document.getElementById("seatbooking");
+//if (iFrame) {
+//    let iFrameDocument = iFrame.contentWindow.document;
+//    iFrame.height = iFrame.contentWindow.document.body.scrollHeight + 20;
+//}
